Moved templates to core::mod
