+++ src/core/masks.rs
@@ -1,9 +1,4 @@
-use super::templates::Square;
-
-
-type SQ = u8;
-type BitBoard = u64;
-
+use super::sq::SQ;
 
 pub const PLAYER_CNT: usize = 2;
 pub const PIECE_CNT: usize = 6;
@@ -72,49 +67,49 @@ pub const NORTH_WEST: i8 = 7;
 pub const SOUTH_EAST: i8 = -7;
 pub const SOUTH_WEST: i8 = -9;
 
-pub const START_BIT_BOARDS: [[BitBoard; PIECE_CNT]; PLAYER_CNT] = [
+pub const START_BIT_BOARDS: [[u64; PIECE_CNT]; PLAYER_CNT] = [
     [START_W_PAWN, START_W_KNIGHT, START_W_BISHOP,
         START_W_ROOK, START_W_QUEEN, START_W_KING, ],
     [START_B_PAWN, START_B_KNIGHT, START_B_BISHOP,
         START_B_ROOK, START_B_QUEEN, START_B_KING, ],
 ];
 
-pub const BLANK_BIT_BOARDS: [[BitBoard; PIECE_CNT]; PLAYER_CNT] =
+pub const BLANK_BIT_BOARDS: [[u64; PIECE_CNT]; PLAYER_CNT] =
     [[0, 0, 0, 0, 0, 0], [0, 0, 0, 0, 0, 0]];
 
-pub const START_OCC_BOARDS: [BitBoard; PLAYER_CNT] = [START_WHITE_OCC, START_BLACK_OCC];
+pub const START_OCC_BOARDS: [u64; PLAYER_CNT] = [START_WHITE_OCC, START_BLACK_OCC];
 
-pub const START_W_PAWN: BitBoard =
+pub const START_W_PAWN: u64 =
     0b00000000_00000000_00000000_00000000_00000000_00000000_11111111_00000000;
-pub const START_W_KNIGHT: BitBoard =
+pub const START_W_KNIGHT: u64 =
     0b00000000_00000000_00000000_00000000_00000000_00000000_00000000_01000010;
-pub const START_W_BISHOP: BitBoard =
+pub const START_W_BISHOP: u64 =
     0b00000000_00000000_00000000_00000000_00000000_00000000_00000000_00100100;
-pub const START_W_ROOK: BitBoard =
+pub const START_W_ROOK: u64 =
     0b00000000_00000000_00000000_00000000_00000000_00000000_00000000_10000001;
-pub const START_W_QUEEN: BitBoard =
+pub const START_W_QUEEN: u64 =
     0b00000000_00000000_00000000_00000000_00000000_00000000_00000000_00001000;
-pub const START_W_KING: BitBoard =
+pub const START_W_KING: u64 =
     0b00000000_00000000_00000000_00000000_00000000_00000000_00000000_00010000;
 
-pub const START_B_PAWN: BitBoard =
+pub const START_B_PAWN: u64 =
     0b00000000_11111111_00000000_00000000_00000000_00000000_00000000_00000000;
-pub const START_B_KNIGHT: BitBoard =
+pub const START_B_KNIGHT: u64 =
     0b01000010_00000000_00000000_00000000_00000000_00000000_00000000_00000000;
-pub const START_B_BISHOP: BitBoard =
+pub const START_B_BISHOP: u64 =
     0b00100100_00000000_00000000_00000000_00000000_00000000_00000000_00000000;
-pub const START_B_ROOK: BitBoard =
+pub const START_B_ROOK: u64 =
     0b10000001_00000000_00000000_00000000_00000000_00000000_00000000_00000000;
-pub const START_B_QUEEN: BitBoard =
+pub const START_B_QUEEN: u64 =
     0b00001000_00000000_00000000_00000000_00000000_00000000_00000000_00000000;
-pub const START_B_KING: BitBoard =
+pub const START_B_KING: u64 =
     0b00010000_00000000_00000000_00000000_00000000_00000000_00000000_00000000;
 
-pub const START_WHITE_OCC: BitBoard =
+pub const START_WHITE_OCC: u64 =
     0b00000000_00000000_00000000_00000000_00000000_00000000_11111111_11111111;
-pub const START_BLACK_OCC: BitBoard =
+pub const START_BLACK_OCC: u64 =
     0b11111111_11111111_00000000_00000000_00000000_00000000_00000000_00000000;
-pub const START_OCC_ALL: BitBoard = START_BLACK_OCC | START_WHITE_OCC;
+pub const START_OCC_ALL: u64 = START_BLACK_OCC | START_WHITE_OCC;
 
 pub const WHITE_KING_START: u8 = 4;
 pub const BLACK_KING_START: u8 = 60;
@@ -136,17 +131,17 @@ pub const CASTLING_ROOK_START: [[u8; CASTLING_SIDES]; PLAYER_CNT] =
     ];
 
 
-pub const CASTLING_PATH_WHITE_K_SIDE: u64 = (1 as u64) << (Square::F1 as u32) |
-    (1 as u64) << (Square::G1 as u32);
-pub const CASTLING_PATH_WHITE_Q_SIDE: u64 = (1 as u64) << (Square::B1 as u32) |
-    (1 as u64) << (Square::C1 as u32) |
-    (1 as u64) << (Square::D1 as u32);
+pub const CASTLING_PATH_WHITE_K_SIDE: u64 = (1 as u64) << SQ::F1.0 as u32 |
+    (1 as u64) << SQ::G1.0 as u32;
+pub const CASTLING_PATH_WHITE_Q_SIDE: u64 = (1 as u64) << SQ::B1.0 as u32 |
+    (1 as u64) << SQ::C1.0 as u32 |
+    (1 as u64) << SQ::D1.0 as u32;
 
-pub const CASTLING_PATH_BLACK_K_SIDE: u64 = (1 as u64) << (Square::F8 as u32) |
-    (1 as u64) << (Square::G8 as u32);
-pub const CASTLING_PATH_BLACK_Q_SIDE: u64 = (1 as u64) << (Square::B8 as u32) |
-    (1 as u64) << (Square::C8 as u32) |
-    (1 as u64) << (Square::D8 as u32);
+pub const CASTLING_PATH_BLACK_K_SIDE: u64 = (1 as u64) << SQ::F8.0 as u32 |
+    (1 as u64) << SQ::G8.0 as u32;
+pub const CASTLING_PATH_BLACK_Q_SIDE: u64 = (1 as u64) << SQ::B8.0 as u32 |
+    (1 as u64) << SQ::C8.0 as u32 |
+    (1 as u64) << SQ::D8.0 as u32;
 
 pub const CASTLING_PATH_WHITE: [u64; CASTLING_SIDES] =
     [CASTLING_PATH_WHITE_K_SIDE, CASTLING_PATH_WHITE_Q_SIDE];
@@ -159,7 +154,7 @@ pub const CASTLING_PATH: [[u64; CASTLING_SIDES]; PLAYER_CNT] =
         [CASTLING_PATH_BLACK_K_SIDE, CASTLING_PATH_BLACK_Q_SIDE],
     ];
 
-pub const SQ_DISPLAY_ORDER: [SQ; SQ_CNT] = [56, 57, 58, 59, 60, 61, 62, 63,
+pub const SQ_DISPLAY_ORDER: [u8; SQ_CNT] = [56, 57, 58, 59, 60, 61, 62, 63,
     48, 49, 50, 51, 52, 53, 54, 55,
     40, 41, 42, 43, 44, 45, 46, 47,
     32, 33, 34, 35, 36, 37, 38, 39,
